requirements.txt
langchain==0.2.10
langchain-openai==0.1.17
python-dotenv==1.0.1
langchain-groq==0.1.6